Added converting strings to basic times and vice versa + read/write bug.
Read + write now accept variables as their parameters for use in functions. Also fixed != precedence bug. Started working on creating a function in pr1.spl that takes the problem input and parses it into a table that can be worked with. Still need to add a way for file data to be piped into the program, as will be done when our program language is tested.